chore: more typecheck

+++ bots/v10/utils/d_star.py
@@ -184,7 +184,7 @@ class DStarLite:
     def plan(self) -> None:
         self._compute_shortest_path()
 
-    def extract_path_lines(self) -> list[tuple[float, float, float, float]]:
+    def extract_path_lines(self) -> list[tuple[int, int, int, int]]:
         """Return line segments [(x1,y1,x2,y2), ...] for rendering."""
         pts = self.extract_path()
         if len(pts) < 2:
@@ -195,7 +195,6 @@ class DStarLite:
             x1, y1 = pts[i]
             x2, y2 = pts[i + 1]
 
-            # center of tiles
             lines.append((x1, y1, x2, y2))
 
         return lines
@@ -330,7 +329,7 @@ class DStarLite:
             return False
         if idx in self._dynamic_blocked:
             return True
-        return (self._block_mask >> self._env._array[idx]) & 1
+        return (self._block_mask >> self._env._array[idx]) & 1 # type: ignore
 
     def _to_idx(self, x: int, y: int) -> int:
         return y * self._w + x
